requirements.txt
matplotlib>=3.3.0
pyyaml>=5.3.1 
opencv-python>=4.6.0   
pillow>=7.1.2   
requests>=2.23.0   
scipy>=1.4.1   
torch>=1.8.0   
torchvision>=0.9.0   
tqdm>=4.64.0    # progress bars
psutil    # system utilization
py-cpuinfo    # display CPU info
thop>=0.1.1    # FLOPs computation
pandas>=1.1.4   
seaborn>=0.11.0  
ipython   
check-manifest   
pre-commit   
pytest   
pytest-cov   
coverage[toml]   
mkdocs-material>=9.5.9   
mkdocstrings[python]   
mkdocs-jupyter    # for notebooks
mkdocs-redirects    # for 301 redirects
mkdocs-ultralytics-plugin>=0.0.43
onnx>=1.12.0   # ONNX export
coremltools>=7.0
openvino-dev>=2023.0
tensorflow

tensorboard
dvclive>=2.12.0
hub-sdk>=0.0.2   # Ultralytics HUBd
ipython   # interactive notebook
albumentations>=1.0.3   # training augmentations
pycocotools>=2.0.7   # COCO mAP